feat: Add Dockerfile for building and running the application in a container
This commit adds a Dockerfile to the repository, allowing the application to be built and run in a containerized environment. The Dockerfile specifies the base image as `golang:1.21.5` and sets up the necessary dependencies for building the application. It also copies the project files, including the `go.mod` and `go.sum` files, and builds the application using the `go build` command. The resulting binary is then copied to a minimal base image, `gcr.io/distroless/base-debian10`, along with the `.env` file. Finally, the Dockerfile sets the command to run the application. This change improves the portability and ease of deployment of the application.

Please note that the commit message provided above is a suggestion. Feel free to modify it according to your project's commit message conventions.